requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
motor==2.5.1
python-decouple==3.6
pytz==2023.3
web3==6.8.0
pycoingecko==3.1.0
pymongo